requirements.txt
streamlit>=1.30.0
google-generativeai>=0.8.0
python-docx>=1.0.0
PyMuPDF>=1.23.0
pytest>=7.0.0
black>=23.0.0
isort>=5.0.0
gTTS>=2.3.0
SpeechRecognition>=3.10.0
reportlab>=4.0.0